+++ src/main/java/org/folio/am/domain/entity/ModuleEntity.java
@@ -20,12 +20,14 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
 import org.folio.common.domain.model.ModuleDescriptor;
+import org.hibernate.annotations.DynamicUpdate;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.annotations.Type;
 import org.hibernate.type.SqlTypes;
 
 @Data
 @Entity
+@DynamicUpdate
 @Table(name = "module")
 @EqualsAndHashCode(callSuper = true)
 public class ModuleEntity extends ArtifactEntity {