+++ src/main/java/org/folio/am/domain/entity/ModuleLocationProjection.java
@@ -1,0 +1,9 @@
+package org.folio.am.domain.entity;
+
+/** Projection for the uncached discovery location of a module. */
+public interface ModuleLocationProjection {
+
+  String getModuleId();
+
+  String getLocation();
+}